# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your super  ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
        "/api/dealer/scrapped-specs": {
            "post": {
                "tags": [
                    "Scrapped Specs"
                ],
                "summary": "get Scrapped Specs List",
                "description": "Returns list",
                "operationId": "scrappedSpecsList",
                "parameters": [
                    {
                        "name": "specs",
                        "in": "query",
                        "required": true
                    },
                    {
                        "name": "category_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "500": {
                        "description": "error"
                    }
                }
            }
        },
        "/api/dealer/store-general-specs/{id}": {
            "post": {
                "tags": [
                    "Store Vehicle General Specs"
                ],
                "summary": "to store Vehicle General Specs",
                "description": "store Vehicle General Specs",
                "operationId": "Store Vehicle General Specs",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "name": "scrapped_spec_id[]",
                        "in": "query",
                        "required": true
                    },
                    {
                        "name": "specs_value[]",
                        "in": "query",
                        "required": true
                    },
                    {
                        "name": "spec_category",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "500": {
                        "description": "error"
                    }
                }
            }
        },
        "/api/dealer/delete-vehicle-specs": {
            "post": {
                "tags": [
                    "delete vehicle specs"
                ],
                "summary": "delete vehicle specs",
                "description": "delete spec list",
                "operationId": "DeleteVehicleSpecs",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "name": "spec_category",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "500": {
                        "description": "error"
                    }
                }
            }
        },
        "/api/show-specs/{slug}": {
            "get": {
                "tags": [
                    "Show vehicle specs"
                ],
                "summary": "Show vehicle specs",
                "description": "Show spec list",
                "operationId": "Show specs",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "500": {
                        "description": "error"
                    }
                }
            }
        }
    }
}